+++ lib/stagehand.dart
@@ -30,9 +30,7 @@ import 'generators/console_simple.dart';
 import 'generators/package_simple.dart';
 import 'generators/server_shelf.dart';
 import 'generators/web_angular.dart';
-import 'generators/web_angular_simple.dart';
 import 'generators/web_simple.dart';
-import 'generators/web_stagexl.dart';
 import 'src/common.dart';
 
 /// A curated, prescriptive list of Dart project generators.
@@ -42,9 +40,7 @@ final List<Generator> generators = [
   new PackageSimpleGenerator(),
   new ServerShelfGenerator(),
   new WebAngularGenerator(),
-  new WebAngularSimpleGenerator(),
   new WebSimpleGenerator(),
-  new WebStageXlGenerator()
 ]..sort();
 
 Generator getGenerator(String id) {