+++ lib/generators/web_simple.dart
@@ -11,8 +11,8 @@ import 'web_simple_data.dart';
  */
 class WebSimpleGenerator extends DefaultGenerator {
   WebSimpleGenerator()
-      : super('web-simple', 'Simple Web Application',
-            'An absolute bare-bones web app.',
+      : super('web-simple', 'Low-level HTML App',
+            'A bare-bones web app.',
             categories: const ['dart', 'web']) {
     for (TemplateFile file in decodeConcatenatedData(data)) {
       addTemplateFile(file);