+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## Unreleased
 
-- Per issue #390, removed the `web-angular-simple` and `web-stagexl` templates.
-  Documentation (such as the
-  [Angular guide](https://webdev.dartlang.org/angular/guide/setup)
+- Simplified what had become a too-long list of templates by removing the
+  `web-angular-simple` and `web-stagexl` templates. Documentation (such
+  as the [Angular guide](https://webdev.dartlang.org/angular/guide/setup))
   that used to rely on `web-angular-simple` now uses
   https://github.com/angular-examples/quickstart.